math/test-fenvinline: Cast fe_exc to unsigned int before printing
On Alpha and IA-64, fexcept_t is unsigned long.  But all the values
fit within an int, so the cast is ok for printing.  All other hosts
use unsigned int or unsigned short already.

+++ math/test-fenvinline.c
@@ -249,7 +249,7 @@ feenable_test (const char *flag_name, fexcept_t fe_exc)
   if (excepts != fe_exc)
     {
       printf ("Test: fedisableexcept (%s) failed, return should be 0x%x, is 0x%x\n",
-              flag_name, fe_exc, excepts);
+              flag_name, (unsigned int)fe_exc, excepts);
       ++count_errors;
     }
 
@@ -284,7 +284,7 @@ feenable_test (const char *flag_name, fexcept_t fe_exc)
   if (excepts != fe_exc)
     {
       printf ("Test: fedisableexcept (%s) failed, return should be 0x%x, is 0x%x\n",
-              flag_name, fe_exc, excepts);
+              flag_name, (unsigned int)fe_exc, excepts);
       ++count_errors;
     }
 #endif
